requirements-rl.txt
torch>=2.0
numpy>=1.24
python-chess>=1.999
